dtype float32

+++ code/models/decoders/mlp.py
@@ -1,6 +1,6 @@
 import torch.nn as nn
 import torch
-torch.set_default_dtype(torch.float64)
+torch.set_default_dtype(torch.float32)
 
 
 class MLP(nn.Module):
